Add a missing defined(OS_BSD). Fix a race in TimeZoneMonitorLinuxImpl().
If TimeZoneMonitorLinuxImpl::StartWatchingOnFileThread() ends up running
before the TimeZoneMonitorLinuxImpl::TimeZoneMonitorLinuxImpl() constructor
finishes, the TimeZoneMonitorLinuxImpl::~TimeZoneMonitorLinuxImpl()
destructor gets called, because it sees a reference count of 0 on the
TimeZoneMonitorLinuxImpl object.

With this change, we do a separate call after constructing the object,
and the reference count of the TimeZoneMonitorLInuxImpl object will be
positive, so StartWatchingOnFileThread() won't accidentally trigger the
destructor.